# Optimize Active Record primary key methods