--- a/activerecord/lib/active_record/attribute_methods/primary_key.rb
+++ b/activerecord/lib/active_record/attribute_methods/primary_key.rb
@@ -17,22 +17,36 @@ module ActiveRecord
 
       # Returns the primary key column's value. If the primary key is composite,
       # returns an array of the primary key column values.
-      def id
-        return _read_attribute(@primary_key) unless @primary_key.is_a?(Array)
-
-        @primary_key.map { |pk| _read_attribute(pk) }
+      if RUBY_VERSION < "3"
+        def id
+          if @_composite_primary_key ||= false
+            @primary_key.map { |pk| _read_attribute(pk) }
+          else
+            _read_attribute(@primary_key)
+          end
+        end
+      else
+        def id
+          if @_composite_primary_key
+            @primary_key.map { |pk| _read_attribute(pk) }
+          else
+            _read_attribute(@primary_key)
+          end
+        end
       end
 
       def primary_key_values_present? # :nodoc:
-        return id.all? if self.class.composite_primary_key?
-
-        !!id
+        if @_composite_primary_key
+          id.all?
+        else
+          !!id
+        end
       end
 
       # Sets the primary key column's value. If the primary key is composite,
       # raises TypeError when the set value not enumerable.
       def id=(value)
-        if self.class.composite_primary_key?
+        if @_composite_primary_key
           raise TypeError, "Expected value matching #{self.class.primary_key.inspect}, got #{value.inspect}." unless value.is_a?(Enumerable)
           @primary_key.zip(value) { |attr, value| _write_attribute(attr, value) }
         else
@@ -43,7 +57,7 @@ module ActiveRecord
       # Queries the primary key column's value. If the primary key is composite,
       # all primary key column values must be queryable.
       def id?
-        if self.class.composite_primary_key?
+        if @_composite_primary_key
           @primary_key.all? { |col| _query_attribute(col) }
         else
           _query_attribute(@primary_key)
@@ -53,7 +67,7 @@ module ActiveRecord
       # Returns the primary key column's value before type cast. If the primary key is composite,
       # returns an array of primary key column values before type cast.
       def id_before_type_cast
-        if self.class.composite_primary_key?
+        if @_composite_primary_key
           @primary_key.map { |col| attribute_before_type_cast(col) }
         else
           attribute_before_type_cast(@primary_key)
@@ -63,7 +77,7 @@ module ActiveRecord
       # Returns the primary key column's previous value. If the primary key is composite,
       # returns an array of primary key column previous values.
       def id_was
-        if self.class.composite_primary_key?
+        if @_composite_primary_key
           @primary_key.map { |col| attribute_was(col) }
         else
           attribute_was(@primary_key)
@@ -73,7 +87,7 @@ module ActiveRecord
       # Returns the primary key column's value from the database. If the primary key is composite,
       # returns an array of primary key column values from database.
       def id_in_database
-        if self.class.composite_primary_key?
+        if @_composite_primary_key
           @primary_key.map { |col| attribute_in_database(col) }
         else
           attribute_in_database(@primary_key)
@@ -81,7 +95,7 @@ module ActiveRecord
       end
 
       def id_for_database # :nodoc:
-        if self.class.composite_primary_key?
+        if @_composite_primary_key
           @primary_key.map { |col| @attributes[col].value_for_database }
         else
           @attributes[@primary_key].value_for_database
@@ -89,6 +103,11 @@ module ActiveRecord
       end
 
       private
+        def init_internals
+          super
+          @_composite_primary_key = self.class.composite_primary_key?
+        end
+
         def attribute_method?(attr_name)
           attr_name == "id" || super
         end