`ActiveRecord::Base#id` became `~10%` slower in 7.1 following the introduction of composite primary keys.

Some production profile revealed a significant amount of time spent in `@primary_key.is_a?(Array)`. Some of it may be the common StackProf bias, but it's still significant.

By storing whether the primary key is composite as a boolean we can reclaim about half of the overhead:

```
ruby 3.2.2 (2023-03-30 revision e51014f9c0) [arm64-darwin22]
Warming up --------------------------------------
       User#id (7.0)   507.375k i/100ms
       User#id (7.1)   460.049k i/100ms
       User#id (opt)   486.487k i/100ms
Calculating -------------------------------------
       User#id (7.0)      5.058M (± 1.1%) i/s -     25.369M in   5.016103s
       User#id (7.1)      4.592M (± 0.8%) i/s -     23.002M in   5.009319s
       User#id (opt)      4.872M (± 0.9%) i/s -     24.811M in   5.093091s

Comparison:
       User#id (7.0):  5058058.8 i/s
       User#id (opt):  4871876.3 i/s - 1.04x  slower
       User#id (7.1):  4592198.3 i/s - 1.10x  slower
```

```
ruby 3.1.4p223 (2023-03-30 revision 957bb7cb81) [arm64-darwin22]
Warming up --------------------------------------
       User#id (7.0)   502.562k i/100ms
       User#id (7.1)   430.043k i/100ms
       User#id (opt)   482.961k i/100ms
Calculating -------------------------------------
       User#id (7.0)      5.060M (± 0.9%) i/s -     25.631M in   5.065498s
       User#id (7.1)      4.325M (± 0.8%) i/s -     21.932M in   5.071700s
       User#id (opt)      4.804M (± 0.8%) i/s -     24.148M in   5.027168s

Comparison:
       User#id (7.0):  5060312.9 i/s
       User#id (opt):  4803851.1 i/s - 1.05x  slower
       User#id (7.1):  4324672.3 i/s - 1.17x  slower
```

<details><summary>Benchmark</summary>

```ruby
begin
  require "bundler/inline"
rescue LoadError => e
  $stderr.puts "Bundler version 1.10 or later is required. Please update your Bundler"
  raise e
end

gemfile(true) do
  source "https://rubygems.org"
  gem "rails", path: "/Users/byroot/src/github.com/Shopify/rails"
  gem "sqlite3"
  gem "benchmark-ips"
end

require "active_record"

ActiveRecord::Base.establish_connection(adapter: "sqlite3", database: ":memory:")
ActiveRecord::Base.logger = Logger.new(STDOUT)

ActiveRecord::Schema.define do
  create_table :users do |t|
    t.string :first_name
    t.string :last_name
  end
end

class User < ActiveRecord::Base
end

module ActiveRecord
  module AttributeMethods
    # = Active Record Attribute Methods Primary Key
    module PrimaryKey
      def initialize(...)
        @composite_primary_key = self.class.composite_primary_key?
        super
      end

      def rails_71_id
        return _read_attribute(@primary_key) unless @primary_key.is_a?(Array)

        @primary_key.map { |pk| _read_attribute(pk) }
      end

      def rails_70_id
        _read_attribute(@primary_key)
      end

      def optimized_id
        if @composite_primary_key
          @primary_key.map { |pk| _read_attribute(pk) }
        else
          _read_attribute(@primary_key)
        end
      end
    end
  end
end

user = User.new

puts RUBY_DESCRIPTION

Benchmark.ips do |x|
  x.report("User#id (7.0)") do
    user.rails_70_id
  end

  x.report("User#id (7.1)") do
    user.rails_71_id
  end

  x.report("User#id (opt)") do
    user.optimized_id
  end

  x.compare!(order: :baseline)
end
```
</details>

Another avenue could be to swap the method entirely. e.g. include a module when you activate composite primary keys on a model. But then it would make it tricky to undo it in a child class. So I'm still on the fence, I'll sleep on it.

cc @ptolts you reported the profiling.
cc @nvasilevski because CPK.