@casperisfine 2023-10-26
> we should look into making the `composite_primary_key?`

You mean on the class? Yeah, it's a good point.