@rafaelfranca 2023-10-26
May I talk about the word of the saviour Strategy Pattern? This is pretty much the text book definition of it. I'm not sure if will be faster than the `@_composite_primary_key` but the code will be simpler to understand and very likely to have similar performance. Let me try.