@rafaelfranca 2023-10-26
I'm running the benchmark script and I keep seeing 7.0 and 7.1 with the same performance. Am I missing something?

```ruby
require "bundler/setup"

require "active_record"
require "benchmark/ips"

ActiveRecord::Base.establish_connection(adapter: "sqlite3", database: ":memory:")
ActiveRecord::Base.logger = Logger.new(STDOUT)

ActiveRecord::Schema.define do
  create_table :users do |t|
    t.string :first_name
    t.string :last_name
  end
end

class User < ActiveRecord::Base
end

module ActiveRecord
  module AttributeMethods
    # = Active Record Attribute Methods Primary Key
    module PrimaryKey
      def rails_71_id
        return _read_attribute(@primary_key) unless @primary_key.is_a?(Array)

        @primary_key.map { |pk| _read_attribute(pk) }
      end

      def rails_70_id
        _read_attribute(@primary_key)
      end
    end
  end
end

user = User.new

puts RUBY_DESCRIPTION

Benchmark.ips do |x|
  x.report("User#id (7.0)") do
    user.rails_70_id
  end

  x.report("User#id (7.1)") do
    user.rails_71_id
  end

  x.compare!(order: :baseline)
end
```

```
ruby 3.2.2 (2023-03-30 revision e51014f9c0) [aarch64-linux]
Warming up --------------------------------------
       User#id (7.0)   494.384k i/100ms
       User#id (7.1)   469.281k i/100ms
Calculating -------------------------------------
       User#id (7.0)      4.907M (± 1.9%) i/s -     24.719M in   5.039524s
       User#id (7.1)      4.509M (± 8.8%) i/s -     22.525M in   5.056944s

Comparison:
       User#id (7.0):  4906874.9 i/s
       User#id (7.1):  4509178.8 i/s - same-ish: difference falls within error
```