@nvasilevski 2023-10-26
@rafaelfranca The difference seems to be `400K i/s` in your case and `~500K i/s` in Jean's case but in your case it's being treated as same-ish. 

This is what I get if I run benchmark from the description:
```
Calculating -------------------------------------
       User#id (7.0)      4.238M (± 6.9%) i/s -     21.081M in   5.003097s
       User#id (7.1)      3.569M (±12.1%) i/s -     17.643M in   5.041735s
       User#id (opt)      4.139M (± 8.6%) i/s -     21.029M in   5.132130s

Comparison:
       User#id (7.0):  4238266.3 i/s
       User#id (opt):  4139040.4 i/s - same-ish: difference falls within error
       User#id (7.1):  3569303.9 i/s - same-ish: difference falls within error
```