@rafaelfranca 2023-10-26
I get all of them falling within error. You also get the same, so maybe it isn't more expensive, neither the optimize version is faster?