@rafaelfranca 2023-10-26
I tried the strategy implementation, it is naive and should be implemented, but I got the same results in the benchmark.

```diff
diff --git a/activerecord/lib/active_record/attribute_methods/primary_key.rb b/activerecord/lib/active_record/attribute_methods/primary_key.rb
index 4adee074e9..ab3949caad 100644
--- a/activerecord/lib/active_record/attribute_methods/primary_key.rb
+++ b/activerecord/lib/active_record/attribute_methods/primary_key.rb
@@ -8,6 +8,34 @@ module AttributeMethods
     module PrimaryKey
       extend ActiveSupport::Concern
 
+      class Strategy
+        def self.build(record, key)
+          if key.is_a?(Array)
+            Composite.new(record, key)
+          else
+            Single.new(record, key)
+          end
+        end
+
+        class Single
+          def initialize(record, key)
+            @record = record
+            @key = key
+          end
+
+          def id
+            @record._read_attribute(@key)
+          end
+        end
+
+        class Composite < Single
+          def id
+            @key.map { |pk| @record._read_attribute(pk) }
+          end
+        end
+      end
+
+
       # Returns this record's primary key value wrapped in an array if one is
       # available.
       def to_key
@@ -18,9 +46,7 @@ def to_key
       # Returns the primary key column's value. If the primary key is composite,
       # returns an array of the primary key column values.
       def id
-        return _read_attribute(@primary_key) unless @primary_key.is_a?(Array)
-
-        @primary_key.map { |pk| _read_attribute(pk) }
+        @primary_key_strategy.id
       end
 
       def primary_key_values_present? # :nodoc:
diff --git a/activerecord/lib/active_record/core.rb b/activerecord/lib/active_record/core.rb
index 93150822c2..3b0ac6c6ad 100644
--- a/activerecord/lib/active_record/core.rb
+++ b/activerecord/lib/active_record/core.rb
@@ -745,7 +745,8 @@ def init_internals
 
         klass = self.class
 
-        @primary_key         = klass.primary_key
+        @primary_key = klass.primary_key
+        @primary_key_strategy = AttributeMethods::PrimaryKey::Strategy.build(self, @primary_key)
         @strict_loading      = klass.strict_loading_by_default
         @strict_loading_mode = :all
 
```

Bechmark

```ruby
require "bundler/setup"

require "active_record"
require "benchmark/ips"

ActiveRecord::Base.establish_connection(adapter: "sqlite3", database: ":memory:")
ActiveRecord::Base.logger = Logger.new(STDOUT)

ActiveRecord::Schema.define do
  create_table :users do |t|
    t.string :first_name
    t.string :last_name
  end
end

class User < ActiveRecord::Base
end

module ActiveRecord
  module AttributeMethods
    # = Active Record Attribute Methods Primary Key
    module PrimaryKey
      def rails_71_id
        return _read_attribute(@primary_key) unless @primary_key.is_a?(Array)

        @primary_key.map { |pk| _read_attribute(pk) }
      end

      def rails_70_id
        _read_attribute(@primary_key)
      end
    end
  end
end

user = User.new

puts RUBY_DESCRIPTION

Benchmark.ips do |x|
  x.report("User#id (7.0)") do
    user.rails_70_id
  end

  x.report("User#id (7.1)") do
    user.rails_71_id
  end

  x.report("User#id (opt)") do
    user.id
  end

  x.compare!(order: :baseline)
end
```