@rafaelfranca 2023-10-26
Ah, something was indeed off. I now get the right results, but I didn't changed anything in the benchmark, and the strategy approach has same performance as the `is_a?` check because it is a few method calls.