@rafaelfranca 2023-10-26
I'm ok with this implementation, it would be ideal to avoid those conditionals or move them all to the same place since they all change at the same time, but the module including is hard to make work and using the strategy pattern is as bad as the `is_a?` check.