@matthewd 2023-10-27
I had definitely pondered the extra-module approach -- have the base implementation stick with the plain key, then when a composite-looking key gets involved, that model pulls in a corresponding module. Given the extra arrays etc, composite mode is already a de-opt, so it's fine to have the module re-check the composite state dynamically... thus no worries about a subclass or later config change un-compositing... the with-module state is "check each time" [what we have now], only the without-module state is specialised [what we had before].