@casperisfine 2023-10-27
> I'm running the benchmark script and I keep seeing 7.0 and 7.1 with the same performance.

So you don't get the same performance, but `benchmark/ips` says the difference is too small for it to be certain it's significant. It's not wrong, but if you run it a few time you see the difference is consistent.

As said previously, if all I got was this benchmark I wouldn't bother, it's really because production profile show it's a big hotspot that I even consider such micro optimizations.